feat(authorize): created custom hook for getting authorized state
